feat(sp-web): reorganize settings into route-driven sections
- split settings into overview, integrations, indexing, account, and about routes

- replace the stubbed About tab with real runtime and project metadata

- extract shared integration repair panel and utilities for setup and settings

- simplify the settings overview to summary content plus section tabs
