+++ apps/sp-web/src/app/pages/settings/settings-account-page.component.html
@@ -1,0 +1,11 @@
+<section class="account-page">
+  <header class="section-head">
+    <div>
+      <p class="eyebrow">Account</p>
+      <h2>Account security</h2>
+    </div>
+    <p class="lede">Manage the single local admin credential used to access Stasharr.</p>
+  </header>
+
+  <app-account-settings-section />
+</section>
