+++ apps/sp-web/src/app/pages/settings/settings-account-page.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+import { AccountSettingsSectionComponent } from '../../features/auth/account-settings-section.component';
+
+@Component({
+  selector: 'app-settings-account-page',
+  imports: [AccountSettingsSectionComponent],
+  templateUrl: './settings-account-page.component.html',
+  styleUrl: './settings-account-page.component.scss',
+})
+export class SettingsAccountPageComponent {}
