+++ apps/sp-web/src/app/shared/integration-repair-panel/integration-repair-panel.component.scss
@@ -1,0 +1,100 @@
+:host {
+  display: block;
+}
+
+.repair-card {
+  border: 1px solid var(--border-subtle);
+  border-radius: 16px;
+  padding: 1rem;
+  background: linear-gradient(180deg, var(--surface-1), var(--surface-2));
+  box-shadow: 0 18px 38px rgba(10, 17, 29, 0.12);
+  display: grid;
+  gap: 0.8rem;
+}
+
+h2,
+p {
+  margin: 0;
+}
+
+.card-head,
+.actions {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 0.75rem;
+}
+
+.eyebrow {
+  font-size: 0.72rem;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  color: var(--text-muted);
+}
+
+.card-summary {
+  font-size: 0.98rem;
+}
+
+.meta,
+.health-meta,
+.action-hint {
+  color: var(--text-muted);
+}
+
+.health-meta,
+form {
+  display: grid;
+  gap: 0.32rem;
+}
+
+form {
+  gap: 0.75rem;
+}
+
+.actions {
+  justify-content: flex-start;
+  flex-wrap: wrap;
+}
+
+.state-badge {
+  display: inline-flex;
+  align-items: center;
+  border-radius: 999px;
+  padding: 0.3rem 0.65rem;
+  font-size: 0.82rem;
+  font-weight: 600;
+  border: 1px solid var(--border-strong);
+  background: var(--surface-elevated);
+  color: var(--text-primary);
+  white-space: nowrap;
+}
+
+.is-ready {
+  border-color: color-mix(in srgb, var(--success-color, #2f855a) 48%, var(--border-strong));
+  background: color-mix(in srgb, var(--success-color, #2f855a) 14%, var(--surface-elevated));
+}
+
+.is-saved {
+  border-color: color-mix(in srgb, var(--warning-color, #b7791f) 48%, var(--border-strong));
+  background: color-mix(in srgb, var(--warning-color, #b7791f) 12%, var(--surface-elevated));
+}
+
+.is-error {
+  border-color: color-mix(in srgb, var(--error-color, #c53030) 52%, var(--border-strong));
+  background: color-mix(in srgb, var(--error-color, #c53030) 12%, var(--surface-elevated));
+}
+
+.is-pending {
+  border-color: var(--border-strong);
+}
+
+@media (max-width: 640px) {
+  .card-head {
+    flex-direction: column;
+  }
+
+  .state-badge {
+    align-self: flex-start;
+  }
+}
